+++ src/main/resources/db/migration/V1_02__create_graphic.sql
@@ -9,7 +9,7 @@ create table graphic
     fill          text,
     fillopacity   text,
     stroke        text,
-    strokeopacity int,
+    strokeopacity float,
     strokewidth   float,
     name          text
 );
