+++ src/main/resources/db/migration/V1_03__create_area.sql
@@ -9,7 +9,7 @@ create table area
     fill                 text,
     fillopacity          float,
     stroke               text,
-    strokeopacity        int,
+    strokeopacity        float,
     strokewidth          float,
     strokelinejoinfill   text,
     strokelinejoinstroke text,
