+++ src/main/resources/db/migration/V1_04__create_symbol.sql
@@ -9,7 +9,7 @@ create table symbol
     fill                 text,
     fillopacity          float,
     stroke               text,
-    strokeopacity        int,
+    strokeopacity        float,
     strokewidth          float,
     stokelinecap         text,
     strokedasharray      text,
