+++ src/main/resources/db/migration/V1_06__create_norm.sql
@@ -9,7 +9,7 @@ create table norm
     fill                 text,
     fillopacity          float,
     stroke               text,
-    strokeopacity        int,
+    strokeopacity        float,
     strokewidth          float,
     strokelinejoinfill   text,
     strokelinejoinstroke text,
